pyproject.toml
[tool.poetry]
name = "sendrag"
version = "0.1.0"
description = "\"A package for semantic retrieval and generative AI chat from SEND related documents\""
authors = ["Md Aminul Islam Prodhan <aminuldu07@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.115.12"
uvicorn = "^0.34.0"
streamlit = "^1.44.1"
faiss-cpu = "^1.10.0"
sentence-transformers = "^4.0.2"
transformers = "4.49.0"
accelerate ="1.6.0"
pypdf2 = "^3.0.1"
openai = "^1.70.0"
torch = "^2.6.0"
flask = "^3.1.0"
numpy = "^2.2.4"
scikit-learn = "^1.6.1"
pymupdf = "^1.26.0"
pycryptodome = "^3.23.0"


[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
pytest = "^8.3.5"
mypy = "^1.15.0"
flake8 = "^7.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
